chatbot_project/requirements.txt
 
flask==3.0.3
requests==2.31.0
python-dotenv